+++ src/sdl_renderer/sdl_renderer.cpp
@@ -41,6 +41,17 @@ SDLRenderer::SDLRenderer(int width, int height, bool fullscreen)
     SetFullScreen(true);
   }
 
+#if defined(__APPLE__)
+  // Apple Silicon Mac + macOS 11.0 だと、
+  // SDL_CreateRenderer をメインスレッドで呼ばないとエラーになる
+  renderer_ = SDL_CreateRenderer(window_, -1, SDL_RENDERER_ACCELERATED);
+  if (renderer_ == nullptr) {
+    RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL_CreateRenderer failed "
+                      << SDL_GetError();
+    return;
+  }
+#endif
+
   thread_ = SDL_CreateThread(SDLRenderer::RenderThreadExec, "Render", this);
 }
 
@@ -50,6 +61,9 @@ SDLRenderer::~SDLRenderer() {
   SDL_WaitThread(thread_, &ret);
   if (ret != 0) {
     RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL Thread error:" << ret;
+  }
+  if (renderer_) {
+    SDL_DestroyRenderer(renderer_);
   }
   if (window_) {
     SDL_DestroyWindow(window_);
@@ -106,12 +120,15 @@ int SDLRenderer::RenderThreadExec(void* data) {
 }
 
 int SDLRenderer::RenderThread() {
+#if !defined(__APPLE__)
   renderer_ = SDL_CreateRenderer(window_, -1, SDL_RENDERER_ACCELERATED);
   if (renderer_ == nullptr) {
     RTC_LOG(LS_ERROR) << __FUNCTION__ << ": SDL_CreateRenderer failed "
                       << SDL_GetError();
     return 1;
   }
+#endif
+
   SDL_SetRenderDrawColor(renderer_, 0, 0, 0, 255);
 
   uint32_t start_time, duration;
@@ -160,6 +177,7 @@ int SDLRenderer::RenderThread() {
   }
 
   SDL_DestroyRenderer(renderer_);
+  renderer_ = nullptr;
 
   return 0;
 }