+++ src/util.cpp
@@ -9,17 +9,13 @@
 #include <boost/beast/version.hpp>
 #include <boost/filesystem/operations.hpp>
 #include <boost/filesystem/path.hpp>
+#include <boost/json.hpp>
 #include <boost/preprocessor/stringize.hpp>
 
 // WebRTC
 #include <rtc_base/helpers.h>
 
-// nlohamnn/json
-#include <nlohmann/json.hpp>
-
 #include "momo_version.h"
-
-using json = nlohmann::json;
 
 void Util::ParseArgs(int argc,
                      char* argv[],
@@ -325,12 +321,12 @@ void Util::ParseArgs(int argc,
 
   auto is_json = CLI::Validator(
       [](std::string input) -> std::string {
-        try {
-          auto _ = json::parse(input);
-          return std::string();
-        } catch (json::parse_error& e) {
+        boost::json::error_code ec;
+        boost::json::parse(input);
+        if (ec) {
           return "Value " + input + " is not JSON Value";
         }
+        return std::string();
       },
       "JSON Value");
   std::string sora_metadata;
@@ -362,7 +358,7 @@ void Util::ParseArgs(int argc,
 
   // メタデータのパース
   if (!sora_metadata.empty()) {
-    args.sora_metadata = json::parse(sora_metadata);
+    args.sora_metadata = boost::json::parse(sora_metadata);
   }
 
   if (args.test_document_root.empty()) {